+++ src/agnocast_cuda/include/agnocast/cuda/message_types.hpp
